move doc computed properties to addon
