+++ addon/properties/index.js
@@ -1,0 +1,5 @@
+import docById from './doc-by-id';
+
+export {
+  docById
+};
